Modified to import and added new call to File Menu
Modified the import to use relative import. Added New File and Merge File to the File Menu. The first allow the user to start over with a clean project without the need to restart the program. The second allow the user to merge the currently open file and a new file, defining the current selected category as the base category for the new data.